fix: fix failing test because of conflicting aria roles
