+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import { AboutSection } from "@/components/sections/about-section";
+import { AboutSection } from "@/components/sections/about/about-section";
 import { ContactSection } from "@/components/sections/contact-section";
 import { HomeSection } from "@/components/sections/home/home-section";
 import { ProjectsSection } from "@/components/sections/projects-section";
