+++ src/components/sections/about/about-section.tsx
@@ -1,0 +1,14 @@
+import { Heading } from "./heading";
+import { Content } from "./content";
+
+export function AboutSection() {
+  return (
+    <div
+      className="min-h-screen w-full flex justify-evenly items-center"
+      id="about"
+    >
+      <Heading />
+      <Content />
+    </div>
+  );
+}
