+++ src/components/sections/about/content.tsx
@@ -1,32 +1,8 @@
-"use client";
-
 import * as React from "react";
-import { ChevronDownIcon } from "@radix-ui/react-icons";
 import { useInView } from "framer-motion";
+import { ChevronDownIcon } from "@radix-ui/react-icons";
 
-const Heading = () => {
-  const ref = React.useRef(null);
-  const isInView = useInView(ref, { once: true });
-
-  return (
-    <div className="pb-4" ref={ref}>
-      <div
-        style={{
-          transform: isInView ? "none" : "translateX(-200px)",
-          opacity: isInView ? 1 : 0,
-          transition: "all 0.9s cubic-bezier(0.17, 0.55, 0.55, 1) 0.5s",
-        }}
-      >
-        <h1 className="font-bold text-5xl md:text-6xl lg:text-8xl">About</h1>
-        <p role="doc-subtitle" className="text-xl">
-          Find out all about who I am and what I do
-        </p>
-      </div>
-    </div>
-  );
-};
-
-const Content = () => {
+export function Content() {
   const ref = React.useRef(null);
   const isInView = useInView(ref, { once: true });
 
@@ -58,18 +34,6 @@ const Content = () => {
           </a>
         </div>
       </div>
-    </div>
-  );
-};
-
-export function AboutSection() {
-  return (
-    <div
-      className="min-h-screen w-full flex justify-evenly items-center"
-      id="about"
-    >
-      <Heading />
-      <Content />
     </div>
   );
 }
